Separación de Clases
Se ha separado las subclases para mantener una estructura mas organizada . También se ha corregido el nombre de la Clase Demonstracio a Demostracio.
