+++ src/dades/Accions/Accio.java
@@ -50,111 +50,5 @@ public abstract class Accio {
     public String getResponsable() {
         return responsable;
     }
-
-
-
    
-}
-
-/**
- * Subclasse demostració.
- */
-class Demonstracio extends Accio {
-    private String dataDisseny; // Data en què es va dissenyar la demostració
-    private boolean activa; // Si la demostració encara és vàlida
-    private int vegadesOfertes; // Nombre de vegades que s'ha ofert
-    private double costMaterials; // Cost per crear els materials de la demostració
-
-    public Demonstracio(String titol, String[] associacions, String responsable, String dataDisseny,
-                        boolean activa, double costMaterials) {
-        super(titol, associacions, responsable);
-        this.dataDisseny = dataDisseny;
-        this.activa = activa;
-        this.vegadesOfertes = 0;
-        this.costMaterials = costMaterials;
-    }
-
-    public String getDataDisseny() {
-        return dataDisseny;
-    }
-
-    public boolean isActiva() {
-        return activa;
-    }
-
-    public int getVegadesOfertes() {
-        return vegadesOfertes;
-    }
-
-    public double getCostMaterials() {
-        return costMaterials;
-    }
-
-    public void incrementarVegadesOfertes() {
-        vegadesOfertes++;
-    }
-
-}
-
-/**
- * Subclasse xerrada.
- */
-class Xerrada extends Accio {
-    private String dataXerrada; // Data en què es farà la xerrada
-    private String[] membresImpartidors; // Membres que imparteixen la xerrada
-    private int assistents; // Nombre d'assistents
-    private int[] valoracions; // Valoracions fetes pels assistents
-    private int numMembres;
-    private int numValoracions;
-
-    public Xerrada(String titol, String[] associacions, String responsable, String dataXerrada) {
-        super(titol, associacions, responsable);
-        this.dataXerrada = dataXerrada;
-        this.membresImpartidors = new String[3]; // Fins a 3 membres
-        this.assistents = 0;
-        this.valoracions = new int[100]; // Fins a 100 valoracions
-        this.numMembres = 0;
-        this.numValoracions = 0;
-    }
-
-    public String getDataXerrada() {
-        return dataXerrada;
-    }
-
-    public String[] getMembresImpartidors() {
-        String[] impartidors = new String[numMembres];
-        System.arraycopy(membresImpartidors, 0, impartidors, 0, numMembres);
-        return impartidors;
-    }
-
-    public int getAssistents() {
-        return assistents;
-    }
-
-    public int[] getValoracions() {
-        int[] numV = new int[numValoracions];
-        System.arraycopy(valoracions, 0, numV, 0, numValoracions);
-        return numV;
-    }
-
-    public void afegirMembreImpartidor(String membre) {
-        if (numMembres < 3) {
-            membresImpartidors[numMembres++] = membre;
-        }
-    }
-
-    public void afegirAssistents(int nombre) {
-        if (nombre >= 0) {
-            assistents += nombre;
-        }
-    }
-
-    public void afegirValoracio(int valoracio) {
-        if (valoracio >= 0 && valoracio <= 10 && numValoracions < valoracions.length) {
-            valoracions[numValoracions++] = valoracio;
-        }
-    }
-
-    
-
 }
